Update environment handling for evaluating block statements
This was initially incorrect. Upon evaluating a block statement, a new environment must be created, with the current environment as the enclosed env. This ensures assignments are handled appropriately, by updating the enclosed variable's state. Doing so means once the enclosed state is returned to the interpreter's current env, the updated vars will carry through and is required for evaluating certain constructs like a for loop's condition expression.