Add break syntax
break statements are only valid within blocks of loops, otherwise the parser will return a syntax error.

nested loops / blocks with multiple breaks only return control to the end of the enclosing block. this means a nested for loop or while loop with a break will return to the end of its block, and will not return from the enclosing for or while loop.